Refactoring into logical sections. Refactoring to use namespaces
