+++ TechTest/AnyCompany/AnyCompany.csproj
@@ -40,11 +40,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Customer.cs" />
-    <Compile Include="CustomerRepository.cs" />
-    <Compile Include="Order.cs" />
-    <Compile Include="OrderRepository.cs" />
-    <Compile Include="OrderService.cs" />
+    <Compile Include="Models\CustomerModel.cs" />
+    <Compile Include="DataRepositories\CustomerRepository.cs" />
+    <Compile Include="Models\OrderModel.cs" />
+    <Compile Include="DataRepositories\OrderRepository.cs" />
+    <Compile Include="Services\OrderService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
