+++ TechTest/AnyCompany/Models/OrderModel.cs
@@ -1,6 +1,6 @@
-﻿namespace AnyCompany
+﻿namespace AnyCompany.Models
 {
-    public class Order
+    public class OrderModel
     {
         public int OrderId { get; set; }
         public double Amount { get; set; }
